+++ dap-etl/extraction/src/test/scala/org/broadinstitute/monster/dap/HLESExtractionPipelineBuilderIntegrationSpec.scala
@@ -2,7 +2,7 @@ package org.broadinstitute.monster.dap
 
 import java.time.{LocalDate, LocalTime, OffsetDateTime, ZoneOffset}
 import better.files.File
-import common.{Args, GetRecords, RedcapRequestGeneratorParams}
+import common.{Args, GetRecords, MockOkWrapper, RedcapMsgGenerator, RedcapRequestGeneratorParams}
 import org.broadinstitute.monster.common.PipelineBuilderSpec
 import upack._
