+++ dap-etl/extraction/src/test/scala/org/broadinstitute/monster/dap/common/MockOkWrapper.scala
@@ -1,11 +1,6 @@
-package org.broadinstitute.monster.dap
+package org.broadinstitute.monster.dap.common
 
 import okhttp3._
-import org.broadinstitute.monster.dap.common.{
-  HttpWrapper,
-  RedCapClient,
-  RedcapRequestGeneratorParams
-}
 import upack.Msg
 
 import scala.concurrent.{Future, Promise}
