+++ dap-etl/extraction/src/test/scala/org/broadinstitute/monster/dap/common/RedCapClientSpec.scala
@@ -1,15 +1,8 @@
-package org.broadinstitute.monster.dap
+package org.broadinstitute.monster.dap.common
 
+import okhttp3.FormBody
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
-import okhttp3.FormBody
-import org.broadinstitute.monster.dap.common.{
-  FilterDirective,
-  FilterOps,
-  GetRecords,
-  RedCapClient,
-  RedcapRequestGeneratorParams
-}
 
 class RedCapClientSpec extends AnyFlatSpec with Matchers {
