+++ dap-etl/extraction/src/test/scala/org/broadinstitute/monster/dap/common/RedcapMsgGenerator.scala
@@ -1,4 +1,4 @@
-package org.broadinstitute.monster.dap
+package org.broadinstitute.monster.dap.common
 
 import upack._
